App_Embrapa/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.1
certifi==2024.2.2
channels==4.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0.4
django-plotly-dash==2.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
dnspython==2.6.1
dpd_components==0.1.0
drf-yasg==1.21.7
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pandasql==0.7.3
parso==0.8.4
platformdirs==4.2.2
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
rich==13.7.1
rpds-py==0.18.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
tornado==6.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
